+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Made the details of holdings excluded from analysis accessible via the activities table
 - Migrated the asset profile dialogs of the market data management in the admin control panel to dedicated routes
 
 ## 3.67.1 - 2026-09-05