+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded `bull-board` from version `9.0.1` to `9.5.0`
 - Upgraded `zod` from version `4.4.3` to `4.5.4`
 
+### Fixed
+
+- Resolved an issue when opening an asset profile from the watchlist
+
 ## 3.68.0 - 2026-09-06
 
 ### Added